+++ tests/test.lunaris.store.offline.js
@@ -1,12 +1,14 @@
 const storeTestUtils     = require('./testUtils');
 const storeOffline       = require('../src/store/store.offline');
-const cache              = require('../src/store/store.cache');
+const cache              = require('../src/cache');
 const collection         = require('../src/store/store.collection');
+const md5                = require('../src/md5');
 
 describe('offline filters', () => {
 
   beforeEach(() => {
     collection.resetVersionNumber();
+    cache.clear();
   });
 
   it('should be defined', () => {
@@ -16,13 +18,12 @@ describe('offline filters', () => {
 
   it('should return an object if the store is a store object', () => {
     var _store = storeTestUtils.initStore('store', {});
-    var _cache = cache.cache();
 
     _store.data.add({
       id : 1
     });
 
-    var _res = storeOffline.filter(_store, _store.data, _cache, { requiredOptions : {}, optionalOptions : {} });
+    var _res = storeOffline.filter(_store, _store.data, { requiredOptions : {}, optionalOptions : {} });
     should(_res).eql({
       id       : 1,
       _rowId   : 1,
@@ -33,13 +34,12 @@ describe('offline filters', () => {
 
   it('should return an object if the store is a store object : no filter values', () => {
     var _store = storeTestUtils.initStore('store', {});
-    var _cache = cache.cache();
 
     _store.data.add({
       id : 1
     });
 
-    var _res = storeOffline.filter(_store, _store.data, _cache, { requiredOptions : {}, optionalOptions : {} });
+    var _res = storeOffline.filter(_store, _store.data, { requiredOptions : {}, optionalOptions : {} });
     should(_res).eql({
       id       : 1,
       _rowId   : 1,
@@ -50,7 +50,6 @@ describe('offline filters', () => {
 
   it('should return an array', () => {
     var _store = storeTestUtils.initStore('store', []);
-    var _cache = cache.cache();
 
     _store.data.add({
       id : 1
@@ -59,7 +58,7 @@ describe('offline filters', () => {
       id : 2
     });
 
-    var _res = storeOffline.filter(_store, _store.data, _cache, { requiredOptions : {}, optionalOptions : {}, cache : { limit : 2, offset : 0 } });
+    var _res = storeOffline.filter(_store, _store.data, { requiredOptions : {}, optionalOptions : {}, cache : { limit : 2, offset : 0 } });
     should(_res).eql([
       {
         id       : 1,
@@ -77,7 +76,6 @@ describe('offline filters', () => {
 
   it('should return an array : no filter values', () => {
     var _store = storeTestUtils.initStore('store', []);
-    var _cache = cache.cache();
 
     _store.data.add({
       id : 1
@@ -86,7 +84,7 @@ describe('offline filters', () => {
       id : 2
     });
 
-    var _res = storeOffline.filter(_store, _store.data, _cache);
+    var _res = storeOffline.filter(_store, _store.data);
     should(_res).eql([
       {
         id       : 1,
@@ -150,7 +148,7 @@ describe('offline filters', () => {
       cache           : { limit : 2, offset : 0 }
     };
 
-    var _filteredValues = storeOffline.filter(_store, _store.data, cache.cache(), _filterValues);
+    var _filteredValues = storeOffline.filter(_store, _store.data, _filterValues);
     should(_filteredValues).be.an.Array().and.have.lengthOf(2);
     should(_filteredValues).eql([
       {
@@ -232,7 +230,7 @@ describe('offline filters', () => {
       cache           : { limit : 2, offset : 0 }
     };
 
-    var _filteredValues = storeOffline.filter(_store, _store.data, cache.cache(), _filterValues);
+    var _filteredValues = storeOffline.filter(_store, _store.data, _filterValues);
     should(_filteredValues).be.an.Array().and.have.lengthOf(1);
     should(_filteredValues).eql([
       {
@@ -294,7 +292,7 @@ describe('offline filters', () => {
       cache : { limit : 2, offset : 0 }
     };
 
-    var _filteredValues = storeOffline.filter(_store, _store.data, cache.cache(), _filterValues);
+    var _filteredValues = storeOffline.filter(_store, _store.data, _filterValues);
     should(_filteredValues).be.an.Array().and.have.lengthOf(2);
     should(_filteredValues).eql([
       {
@@ -313,7 +311,7 @@ describe('offline filters', () => {
     ]);
 
     _filterValues.optionalOptions[0][1] = 'pomme chaude';
-    _filteredValues                     = storeOffline.filter(_store, _store.data, cache.cache(), _filterValues);
+    _filteredValues                     = storeOffline.filter(_store, _store.data, _filterValues);
     should(_filteredValues).be.an.Array().and.have.lengthOf(1);
     should(_filteredValues).eql([
       {
@@ -362,7 +360,7 @@ describe('offline filters', () => {
       cache : { limit : 2, offset : 0 }
     };
 
-    var _filteredValues = storeOffline.filter(_store, _store.data, cache.cache(), _filterValues);
+    var _filteredValues = storeOffline.filter(_store, _store.data, _filterValues);
     should(_filteredValues).be.an.Array().and.have.lengthOf(2);
     should(_filteredValues).eql([
       {
@@ -417,7 +415,7 @@ describe('offline filters', () => {
       cache : { limit : 2, offset : 0 }
     };
 
-    var _filteredValues = storeOffline.filter(_store, _store.data, cache.cache(), _filterValues);
+    var _filteredValues = storeOffline.filter(_store, _store.data, _filterValues);
     should(_filteredValues).be.an.Array().and.have.lengthOf(2);
     should(_filteredValues).eql([
       {
@@ -472,7 +470,7 @@ describe('offline filters', () => {
       cache : { limit : 2, offset : 0 }
     };
 
-    var _filteredValues = storeOffline.filter(_store, _store.data, cache.cache(), _filterValues);
+    var _filteredValues = storeOffline.filter(_store, _store.data, _filterValues);
     should(_filteredValues).be.an.Array().and.have.lengthOf(2);
     should(_filteredValues).eql([
       {
@@ -555,7 +553,7 @@ describe('offline filters', () => {
       cache : { limit : 2, offset : 0 }
     };
 
-    var _filteredValues = storeOffline.filter(_store, _store.data, cache.cache(), _filterValues);
+    var _filteredValues = storeOffline.filter(_store, _store.data, _filterValues);
     should(_filteredValues).be.an.Array().and.have.lengthOf(1);
     should(_filteredValues).eql([
       {
@@ -617,7 +615,7 @@ describe('offline filters', () => {
       cache : { limit : 4, offset : 0 }
     };
 
-    var _filteredValues = storeOffline.filter(_store, _store.data, cache.cache(), _filterValues);
+    var _filteredValues = storeOffline.filter(_store, _store.data, _filterValues);
     should(_filteredValues).be.an.Array().and.have.lengthOf(4);
     should(_filteredValues).eql([
       {
@@ -698,11 +696,11 @@ describe('offline filters', () => {
         limit  : 2,
         offset : 0,
         0      : ['B', 'C']
-      }
+      },
+      constructedRequiredOptions : ''
     };
 
-    var _cache          = cache.cache();
-    var _filteredValues = storeOffline.filter(_store, _store.data, _cache, _filterValues);
+    var _filteredValues = storeOffline.filter(_store, _store.data, _filterValues);
     should(_filteredValues).be.an.Array().and.have.lengthOf(2);
     should(_filteredValues).eql([
       {
@@ -719,23 +717,17 @@ describe('offline filters', () => {
         _version : [4]
       }
     ]);
-    should(_cache.cache()).eql([
-      [
-        {
-          0      : ['B', 'C'],
-          limit  : 2,
-          offset : 0
-        },
-        [3, 4]
-      ],
-      [
-        {
-          0      : ['B', 'C'],
-          limit  : 2,
-          offset : 2
-        },
-        [5, 6]
-      ]
+    should(cache._cache()).eql([
+      {
+        hash   : md5('/store?limit=2&offset=0&search=label' + encodeURIComponent(':[B,C]')),
+        ids    : [3, 4],
+        stores : ['store']
+      },
+      {
+        hash   : md5('/store?limit=2&offset=2&search=label' + encodeURIComponent(':[B,C]')),
+        ids    : [5, 6],
+        stores : ['store']
+      }
     ]);
   });
 
@@ -786,11 +778,11 @@ describe('offline filters', () => {
         limit  : 2,
         offset : 0,
         0      : ['B', 'C']
-      }
+      },
+      constructedRequiredOptions : ''
     };
 
-    var _cache          = cache.cache();
-    var _filteredValues = storeOffline.filter(_store, _store.data, _cache, _filterValues);
+    var _filteredValues = storeOffline.filter(_store, _store.data, _filterValues);
     should(_filteredValues).be.an.Array().and.have.lengthOf(2);
     should(_filteredValues).eql([
       {
@@ -808,23 +800,17 @@ describe('offline filters', () => {
       }
     ]);
 
-    should(_cache.cache()).eql([
-      [
-        {
-          0      : ['B', 'C'],
-          limit  : 2,
-          offset : 0
-        },
-        [3, 4]
-      ],
-      [
-        {
-          0      : ['B', 'C'],
-          limit  : 2,
-          offset : 2
-        },
-        [5]
-      ]
+    should(cache._cache()).eql([
+      {
+        hash   : md5('/store?limit=2&offset=0&search=label' + encodeURIComponent(':[B,C]')),
+        ids    : [3, 4],
+        stores : ['store']
+      },
+      {
+        hash   : md5('/store?limit=2&offset=2&search=label' + encodeURIComponent(':[B,C]')),
+        ids    : [5],
+        stores : ['store']
+      }
     ]);
   });
 
@@ -881,8 +867,7 @@ describe('offline filters', () => {
       }
     };
 
-    var _cache          = cache.cache();
-    var _filteredValues = storeOffline.filter(_store, _store.data, _cache, _filterValues);
+    var _filteredValues = storeOffline.filter(_store, _store.data, _filterValues);
     should(_filteredValues).be.an.Array().and.have.lengthOf(4);
     should(_filteredValues).eql([
       {
@@ -912,15 +897,10 @@ describe('offline filters', () => {
       }
     ]);
 
-    should(_cache.cache()).eql([
-      [
-        {
-          0      : ['B', 'C'],
-          limit  : 4,
-          offset : 0
-        },
-        [3, 4, 5, 6]
-      ]
-    ]);
+    should(cache._cache()).eql([{
+      hash   : md5('/store?limit=4&offset=0&search=label' + encodeURIComponent(':[B,C]')),
+      ids    : [3, 4, 5, 6],
+      stores : ['store']
+    }]);
   });
 });